Update transforms.bbox3droi
This method returns a tensor. There was a condition that led to returns of tensors with 7 or 8 size on dim 1 depending on the fulfillment of the condition. The change fixes the shape to 8 without any other changes. This is necessary because the downstream modules that use this output expect a size of 8 on dim 1 not 7. See `Single3DRoIPointExtractor.forward`